backend/requirements.txt
bson==0.5.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-core==2.2.2
google-cloud-datastore==2.4.0
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
idna==3.3
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
libcst==0.4.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
proto-plus==1.19.9
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
typing-extensions==4.0.1
typing-inspect==0.7.1
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
